Cleaning Data Editor Legacy Code
- Removes lingering variables & functionality due to the rapid
  development of the data editor.
- Fixes issue where the data view content control showed index of
  viewport container instead of file offset.

Closes #771
Closes #773
