+++ src/svelte/src/components/DataDisplays/CustomByteDisplay/BinaryData.ts
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-import { writable, derived } from 'svelte/store'
 import { SimpleWritable } from '../../../stores/localStore'
-import type { BytesPerRow, RadixValues } from '../../../stores/configuration'
+import type { RadixValues } from '../../../stores/configuration'
 import { radixBytePad } from '../../../utilities/display'
 
 export const BYTE_ACTION_DIV_OFFSET: number = 24
@@ -53,17 +52,6 @@ export interface EditByteEvent extends EditEvent {
   targetByte: ByteValue
 }
 
-export enum ViewportBoundaryTrigger {
-  SCROLL_TOP,
-  SCROLL_BOTTOM,
-}
-
-export const scroll_boundary_event = (top: boolean, end: boolean) => {
-  return top
-    ? ViewportBoundaryTrigger.SCROLL_TOP
-    : ViewportBoundaryTrigger.SCROLL_BOTTOM
-}
-
 export const byte_value_string = (value: number, radix: RadixValues) => {
   if (value > 255)
     throw `Value {${value}} is larger than an unsigned int (255).`
@@ -86,8 +74,6 @@ export const RADIX_REGEX_MATCH_STR = {
   8: /[0-8]{3}/g,
   2: /[0-1]{8}/g,
 }
-
-export const processingViewportRefresh = writable(false)
 
 export class ViewportData_t {
   data = new Uint8Array(0)
@@ -160,98 +146,6 @@ export class ViewportDataStore_t extends SimpleWritable<ViewportData_t> {
   }
 }
 
-export const viewport = new ViewportDataStore_t()
-
-export type ByteActionPxOffsets = {
-  insertBefore: {
-    left: number
-    top: number
-  }
-  insertAfter: {
-    left: number
-    top: number
-  }
-  delete: {
-    left: number
-    top: number
-  }
-  input: {
-    left: number
-    top: number
-  }
-}
-
 export function latin1Undefined(charCode: number): boolean {
   return charCode < 32 || (charCode > 126 && charCode < 160)
-}
-
-export function update_byte_action_offsets(
-  targetDiv: HTMLDivElement,
-  offsetTopBy: number = 0,
-  offsetLeftBy: number = 0
-) {
-  const targetWidth = parseInt(targetDiv.style.width.replace('px', '')) + 4
-  byteActionPxOffsets.update((currentOffsets) => {
-    currentOffsets.delete = {
-      left: targetDiv.offsetLeft + offsetLeftBy,
-      top: targetDiv.offsetTop + BYTE_ACTION_DIV_OFFSET - offsetTopBy,
-    }
-    currentOffsets.input = {
-      left: targetDiv.offsetLeft + offsetLeftBy,
-      top: targetDiv.offsetTop - offsetTopBy,
-    }
-    currentOffsets.insertAfter = {
-      left: targetDiv.offsetLeft + targetWidth + offsetLeftBy,
-      top: targetDiv.offsetTop - offsetTopBy,
-    }
-    currentOffsets.insertBefore = {
-      left: targetDiv.offsetLeft - targetWidth + offsetLeftBy,
-      top: targetDiv.offsetTop - offsetTopBy,
-    }
-
-    return currentOffsets
-  })
-}
-
-export enum ByteValuePxWidths {
-  DISPLAY = 20,
-  EDITING = 68,
-}
-
-export let ByteValueArray: Array<ByteValue> = []
-
-export const bytesPerRow = writable(16 as BytesPerRow)
-export const editingByte = writable(false)
-export const selectedByte = writable({
-  text: '',
-  offset: -1,
-  value: -1,
-} as ByteValue)
-export const byteActionPxOffsets = writable({
-  insertBefore: {
-    left: 0,
-    top: 0,
-  },
-  insertAfter: {
-    left: 0,
-    top: 0,
-  },
-  delete: {
-    left: 0,
-    top: 0,
-  },
-  input: {
-    left: 0,
-    top: 0,
-  },
-} as ByteActionPxOffsets)
-export const mouseSelectionBytes = writable({ mousedown: -1, mouseup: -1 })
-
-// derived readable number whose value is the computed number of bytes in the edited file
-export const offsetMax = derived(viewport, ($viewport) => {
-  // this should be the same as the computed file size
-  return $viewport.fileOffset + $viewport.length + $viewport.bytesLeft
-})
-export function focus_byte_input() {
-  document.getElementById('byte-input').focus()
 }
