+++ src/svelte/src/components/DataDisplays/DataViewports.svelte
@@ -18,84 +18,12 @@ limitations under the License.
   import {
     addressRadix,
     bytesPerRow,
-    editMode,
-    editedDataSegment,
-    editorEncoding,
-    selectionDataStore,
-    selectionSize,
-    viewportClientHeight,
-    viewportLineHeight,
-    viewportScrollHeight,
-    viewportScrollTop,
-    viewportLength,
     displayRadix,
     dataFeedLineTop,
     dataFeedAwaitRefresh,
+    viewport,
   } from '../../stores'
-  import {
-    viewport_references,
-    type ViewportReferences,
-  } from '../../utilities/display'
-  import { MessageCommand } from '../../utilities/message'
-  import { onMount, tick } from 'svelte'
-  import { vscode } from '../../utilities/vscode'
-  import { EditByteModes } from '../../stores/configuration'
-  import { viewport } from './CustomByteDisplay/BinaryData'
   import DataLineFeed from './CustomByteDisplay/DataLineFeed.svelte'
-
-  const viewportRefs = viewport_references() as ViewportReferences
-
-  //
-  // reactive statements
-  //
-
-  $: {
-    $editMode === EditByteModes.Single
-      ? postEditorOnChangeMsg('hex')
-      : postEditorOnChangeMsg($editorEncoding)
-
-    // when the viewport length changes, update the viewport geometry
-    if ($viewportLength >= 0) {
-      // populateViewportGeometry()
-    }
-  }
-
-  function populateViewportGeometry() {
-    // event handlers expect synchronous functions, so wrap the async function in a sync function
-    async function populateViewportGeometryOps_() {
-      if (viewportRefs.physical) {
-        // wait for the DOM to be updated before getting the viewport geometry
-        await tick()
-        $viewportScrollTop = viewportRefs.physical.scrollTop
-        $viewportScrollHeight = viewportRefs.physical.scrollHeight
-        $viewportClientHeight = viewportRefs.physical.clientHeight
-        $viewportLineHeight = parseFloat(
-          getComputedStyle(viewportRefs.physical).lineHeight
-        )
-      }
-    }
-
-    populateViewportGeometryOps_()
-  }
-
-  function postEditorOnChangeMsg(forcedEncoding?: string) {
-    vscode.postMessage({
-      command: MessageCommand.editorOnChange,
-      data: {
-        fileOffset: $selectionDataStore.startOffset,
-        selectionData: $editedDataSegment,
-        encoding: forcedEncoding ? forcedEncoding : $editorEncoding,
-        selectionSize: $selectionSize,
-        editMode: $editMode,
-      },
-    })
-  }
-
-  onMount(() => {
-    // populateViewportGeometry()
-    // recalculate the viewport geometry when the window is resized
-    // window.addEventListener('resize', populateViewportGeometry)
-  })
 </script>
 
 <DataLineFeed
