+++ src/svelte/src/components/DataDisplays/Fieldsets/DataView.svelte
@@ -30,12 +30,12 @@ limitations under the License.
     dvUint8,
     selectionDataStore,
     editedDataSegment,
+    selectedByte,
   } from '../../../stores'
   import { ENDIANNESS_OPTIONS } from '../../../stores/configuration'
   import { UIThemeCSSClass } from '../../../utilities/colorScheme'
-  import Input from '../../Inputs/Input/Input.svelte'
   import { createEventDispatcher } from 'svelte'
-  import { selectedByte } from '../CustomByteDisplay/BinaryData'
+  import Input from '../../Inputs/Input/Input.svelte'
 
   const eventDispatcher = createEventDispatcher()
   const ERROR_MESSAGE_TIMEOUT = 5000
