+++ src/svelte/src/components/layouts/Tooltip.svelte
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 -->
 <script lang="ts">
-  import { tooltipsEnabled } from '../../utilities/display'
+  import { tooltipsEnabled } from '../../stores'
 
   const NULL = () => {}
