+++ src/svelte/src/stores/index.ts
@@ -16,28 +16,32 @@
  */
 
 import type { ValidationResponse } from '../utilities/display'
+import { ThemeType } from '../utilities/colorScheme'
+import { FileMetrics } from '../components/Header/fieldsets/FileMetrics'
+import { derived, writable } from 'svelte/store'
+import { SimpleWritable } from './localStore'
+import { ErrorComponentType, ErrorStore } from '../components/Error/Error'
+import {
+  radixBytePad,
+  regexEditDataTest,
+  validateEncodingStr,
+} from '../utilities/display'
+import {
+  BYTE_ACTION_DIV_OFFSET,
+  ViewportDataStore_t,
+  type ByteValue,
+} from '../components/DataDisplays/CustomByteDisplay/BinaryData'
+import {
+  ReplaceQuery,
+  SearchQuery,
+} from '../components/Header/fieldsets/SearchReplace'
 import {
   EditByteModes,
   UNPRINTABLE_CHAR_STAND_IN,
   type RadixValues,
   type BytesPerRow,
   EditActionRestrictions,
-  editorActionsAllowed,
 } from './configuration'
-import { ThemeType } from '../utilities/colorScheme'
-import { fileMetrics } from '../components/Header/fieldsets/FileMetrics'
-import {
-  radixBytePad,
-  regexEditDataTest,
-  validateEncodingStr,
-} from '../utilities/display'
-import { derived, writable } from 'svelte/store'
-import { SimpleWritable } from './localStore'
-import {
-  BYTE_ACTION_DIV_OFFSET,
-  selectedByte,
-  viewport,
-} from '../components/DataDisplays/CustomByteDisplay/BinaryData'
 
 export class SelectionData_t {
   startOffset = -1
@@ -107,21 +111,79 @@ export const searchCaseInsensitive = writable(false)
 export const dataFeedLineTop = writable(0)
 export const dataFeedAwaitRefresh = writable(false)
 export const rerenderActionElements = writable(false)
-// data in the viewport
-// export const viewportData = writable(new Uint8Array(0))
 
 // Viewport properties
-export const viewportStartOffset = writable(0)
-export const viewportLength = writable(0)
-export const viewportFollowingByteCount = writable(0)
-export const viewportScrollTop = writable(0)
-export const viewportScrollHeight = writable(0)
-export const viewportClientHeight = writable(0)
-export const viewportCapacity = writable(0)
-export const viewportLineHeight = writable(0)
+export const viewport = new ViewportDataStore_t()
+export const viewportNumLinesDisplayed = writable(20)
+
+export const bytesPerRow = writable(16 as BytesPerRow)
+export const editingByte = writable(false)
+export const selectedByte = writable({
+  text: '',
+  offset: -1,
+  value: -1,
+} as ByteValue)
+
+export const fileMetrics = new FileMetrics()
+
+export const searchQuery = new SearchQuery()
+export const replaceQuery = new ReplaceQuery()
+
+export const searchErr = new ErrorStore(ErrorComponentType.SYMBOL)
+export const replaceErr = new ErrorStore(ErrorComponentType.SYMBOL)
+export const seekErr = new ErrorStore(ErrorComponentType.SYMBOL)
+
+export const editorActionsAllowed = writable(EditActionRestrictions.None)
+export const tooltipsEnabled = writable(false)
+export const sizeHumanReadable = writable(false)
 
 // tracks the start and end offsets of the current selection
 export const selectionDataStore = new SelectionData()
+
+export const searchable = derived(
+  [searchQuery, editorEncoding],
+  ([$searchQuery, $editorEncoding]) => {
+    if ($searchQuery.input.length === 0 || $searchQuery.processing) {
+      searchErr.update(() => {
+        return ''
+      })
+      return false
+    }
+    const ret = validateEncodingStr($searchQuery.input, $editorEncoding, 'full')
+    searchErr.update(() => {
+      return ret.errMsg
+    })
+    return ret.valid
+  }
+)
+
+export const replaceable = derived(
+  [replaceQuery, editorEncoding, searchable, selectionDataStore],
+  ([$replaceData, $editorEncoding, $searchable, $selectionData]) => {
+    if (
+      $replaceData.input.length < 0 ||
+      !$searchable ||
+      $replaceData.processing
+    ) {
+      replaceErr.update(() => {
+        return ''
+      })
+      return false
+    }
+    if ($selectionData.active) {
+      replaceErr.update(() => {
+        return 'Cannot replace while viewport data is selected'
+      })
+      return false
+    }
+
+    const ret = validateEncodingStr($replaceData.input, $editorEncoding)
+    replaceErr.update(() => {
+      return ret.errMsg
+    })
+    return ret.valid
+  }
+)
 
 // derived readable enumeration that indicates the edit mode (single byte or multiple bytes)
 export const editMode = derived(
@@ -132,41 +194,6 @@ export const editMode = derived(
       : EditByteModes.Multiple
   },
   EditByteModes.Single
-)
-
-// derived readable number whose value is the number of lines displayed in the viewport
-export const viewportNumLinesDisplayed = derived(
-  [viewportClientHeight, viewportLineHeight],
-  ([$viewportClientHeight, $viewportLineHeight]) => {
-    return Math.floor($viewportClientHeight / $viewportLineHeight) + 1
-  }
-)
-
-// derived readable number whose value is the end offset of the current viewport
-export const viewportEndOffset = derived(
-  [viewportStartOffset, viewportLength],
-  ([$viewportStartOffset, $viewportLength]) => {
-    return $viewportStartOffset + $viewportLength
-  }
-)
-
-// derived readable boolean that indicates if the viewport is scrolled to the top
-export const viewportScrolledToTop = derived(
-  [viewportScrollTop],
-  ([$viewportScrollTop]) => {
-    return $viewportScrollTop === 0
-  }
-)
-
-// derived readable boolean that indicates if the viewport is scrolled to the end
-export const viewportScrolledToEnd = derived(
-  [viewportScrollTop, viewportScrollHeight, viewportClientHeight],
-  ([$viewportScrollTop, $viewportScrollHeight, $viewportClientHeight]) => {
-    return (
-      Math.ceil($viewportScrollTop) + $viewportClientHeight >=
-      $viewportScrollHeight
-    )
-  }
 )
 
 // derived readable number whose value is the size of the current data selection
@@ -215,11 +242,6 @@ export const editedByteIsOriginalByte = derived(
       : $editorSelection.toLowerCase() === $selectedByte.text.toLowerCase()
   }
 )
-
-// derived readable number that indicates the number of encoded bytes per row in each viewport
-export const bytesPerRow = derived(displayRadix, ($displayRadix) => {
-  return $displayRadix === 2 ? 8 : (16 as BytesPerRow)
-})
 
 export const viewportColumnWidth = derived(bytesPerRow, (bytesPerRow) => {
   return bytesPerRow * BYTE_ACTION_DIV_OFFSET
@@ -376,10 +398,12 @@ function validRequestableData(
 }
 
 export const dvOffset = derived(
-  [selectionDataStore, addressRadix],
-  ([$selectionData, $addressRadix]) => {
+  [selectionDataStore, addressRadix, viewport],
+  ([$selectionData, $addressRadix, $viewport]) => {
     return $selectionData.active
-      ? $selectionData.startOffset.toString($addressRadix).toUpperCase()
+      ? ($viewport.fileOffset + $selectionData.startOffset)
+          .toString($addressRadix)
+          .toUpperCase()
       : ''
   }
 )
