+++ src/svelte/src/utilities/highlights.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { derived, get, readable, writable } from 'svelte/store'
+import { derived, readable, writable } from 'svelte/store'
 import { selectionDataStore } from '../stores'
 
 let selectionHighlightLUT = new Uint8Array(1024)
